examples: move test into the tti_example.py script
